Add Storage preference (1/2)
This patch allows the user to change the default storage path.

As in CM11, when the path is changed it sends an intent
to the Gallery to set the default camera folder.

Change-Id: I319decc3fdceca61050cf64f7e59afb06fedbf80

Fix writing to external storage.

Change-Id: Ie0917b6c17129cd9ad2c69a2a53a4c3c3ee565b6

CameraSettingsActivity: Handle NPE when storage preference is already removed

Fixes: http://pastebin.com/zv1sPid2

Change-Id: I05951f93b37759f4d8c9f2fde10c84eb388c4830

Forward-port to AOSP Camera2

Change-Id: I544247320b4359418396a16c2f52191ae832fdf4
Signed-off-by: Adrian DC <radian.dc@gmail.com>
Signed-off-by: Harsh Shandilya <harsh@prjkt.io>